+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 
+## v2.5.-
+
+- [FIX] Switch status updates should now work correctly.
+- [REMOVE] Removed custom logger and switched to native homebridge logging interface
 ## v2.4.1
 
 - [NEW] Finally Added support for buttons/remotes under HomeKit.